Fix: external storage problem NavRing and NavBar applications
If the user binds an app which is located on the
external sdcard and is rebooting the button icon
will be left blank because the navring/navbar is already
constructed before the sdcard is mounted. To avoid
this we listen to the app available intents.
If it is not available the ring target/navbar button shows the none icon.
 When available it shows the
correct app icon. So that the user can clearly see
this app is accessable or not. As well when the
user goes into mass storage mode or does
unmount the sdcard or whatever case could
happen that the application is not there.

Reference:
https://review.slimroms.net:8443/#/c/12364/
https://review.slimroms.net:8443/#/c/12363/

Change-Id: Ib5ee29f1b30cff50d4c64f059259666a7618c090