fb: system wide immersive mode aka expanded desktop thanks to
Danny and Kiril for the base

and kroz for the graphics

At least it is expanded desktop mode like on 4.3 and before but using
not the nice new immersive mode from google which was introduced with 4.4
Means everwhere it is swipable in from the top or bottom depends on
if the navbar or statusbar is existend. As well it respects our other
features like 0dp settings for navbar in landscape or portrait etc.
The option to choose the mode what the user want to hide we removed
due that it is now absolutly not needed. 1. it is swipeable now
2. The user can disable navbar like he want in navbar settings if
urgently needed

As well added shortcut support for navbar navring and powermenu

Change-Id: I37e8f9571cf622154d347b7eabd13bffc99ae187

Conflicts:
	core/java/android/provider/Settings.java
	core/res/res/values/slim_strings.xml
	core/res/res/values/slim_symbols.xml
	policy/src/com/android/internal/policy/impl/PhoneWindowManager.java